requirements.txt
cmake
wheel
numpy>1.2
hydra-core==1.1.1
hydra-colorlog
matplotlib
opencv-python
omegaconf
kaggle
plotly
ftfy
pyhash
pytorch-lightning==1.5.9
torch==1.10.1
torchvision
MulticoreTSNE
gitpython
scipy
sentence-transformers
pytorch3d
gym
moviepy
tqdm
termcolor
wandb
pypng
segmentation-models-pytorch
